LAlala

+++ CCharacter.hpp
@@ -22,10 +22,10 @@ private:
 public:
 	CCharacter();
 	~CCharacter();
-	int getId(){return id;}
-	int getX(){player.x;}
-	int getY(){player.y;}
-	int getVelocity(){return velocity;}
+	int16_t getId(){return id;}
+	int16_t getX(){player.x;}
+	int16_t getY(){player.y;}
+	int16_t getVelocity(){return velocity;}
 	Team getTeam(){return (Team)team;}
 	Direction getDir(){return (Direction)dir;}
 	void setDir(Direction dir){this->dir=dir;}
@@ -33,8 +33,8 @@ public:
 	
 	void draw(int x,int y);
 	
-	void setX(int x){this->player.x=x;}
-	void setY(int y){this->player.y=y;}
+	void setX(int16_t x){this->player.x=x;}
+	void setY(int16_t y){this->player.y=y;}
 	void setTeam(Team team){this->team=(int16_t)team;}
 };
